# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FFG Foods e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  /** Price in kobo (1 Naira = 100 kobo) */
  priceKobo: number;
  /** original | hibiscus | ginger-lemon | cinnamon-lemon */
  flavor: string;
  /** sachet | box */
  type: string;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  stockCount: number;
  featured: boolean;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  priceKobo: number;
  flavor: string;
  type: string;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  stockCount: number;
  featured: boolean;
  /** @nullable */
  categoryId?: number | null;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  priceKobo?: number;
  flavor?: string;
  type?: string;
  /** @nullable */
  imageUrl?: string | null;
  inStock?: boolean;
  stockCount?: number;
  featured?: boolean;
  /** @nullable */
  categoryId?: number | null;
}

export interface OrderItem {
  productId: number;
  productName: string;
  productType: string;
  quantity: number;
  priceKobo: number;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress: string;
  /** @nullable */
  deliveryState?: string | null;
  items: OrderItem[];
  subtotalKobo: number;
  deliveryFeeKobo: number;
  totalKobo: number;
  /** pending | processing | shipped | delivered | cancelled */
  status: string;
  /** @nullable */
  notes?: string | null;
  /**
     * Paystack payment reference
     * @nullable
     */
  paymentReference?: string | null;
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress: string;
  /** Nigerian state for delivery fee calculation */
  deliveryState: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paymentReference?: string | null;
  items: OrderItemInput[];
}

export interface OrderStatusUpdate {
  status: string;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface AdminStats {
  totalOrders: number;
  totalRevenueKobo: number;
  pendingOrders: number;
  totalProducts: number;
  recentOrders: Order[];
  ordersByStatus: StatusCount[];
}

export interface SiteSettings {
  heroTitle: string;
  heroSubtitle: string;
  heroCtaText: string;
  heroCtaSecondaryText: string;
  heroImageUrl: string;
  heroSecondaryImageUrl: string;
  companyTagline: string;
  aboutText: string;
  contactPhone: string;
  contactEmail: string;
  contactAddress: string;
  whatsappNumber: string;
  footerText: string;
  announcementBanner: string;
  contactHeroSubtitle: string;
  contactInfoDesc: string;
  contactBusinessHours: string;
  termsLastUpdated: string;
  termsS1: string;
  termsS2: string;
  termsS3: string;
  termsS4: string;
  termsS5: string;
  termsS6: string;
  termsS7: string;
  termsS8: string;
  termsS9: string;
  termsS10: string;
  termsS11: string;
  termsS12: string;
  privacyLastUpdated: string;
  privacyS1: string;
  privacyS2: string;
  privacyS3: string;
  privacyS4: string;
  privacyS5: string;
  privacyS6: string;
  privacyS7: string;
  privacyS8: string;
  privacyS9: string;
  privacyS10: string;
  privacyS11: string;
}

export interface SiteSettingsUpdate {
  heroTitle?: string;
  heroSubtitle?: string;
  heroCtaText?: string;
  heroCtaSecondaryText?: string;
  heroImageUrl?: string;
  heroSecondaryImageUrl?: string;
  companyTagline?: string;
  aboutText?: string;
  contactPhone?: string;
  contactEmail?: string;
  contactAddress?: string;
  whatsappNumber?: string;
  footerText?: string;
  announcementBanner?: string;
  contactHeroSubtitle?: string;
  contactInfoDesc?: string;
  contactBusinessHours?: string;
  termsLastUpdated?: string;
  termsS1?: string;
  termsS2?: string;
  termsS3?: string;
  termsS4?: string;
  termsS5?: string;
  termsS6?: string;
  termsS7?: string;
  termsS8?: string;
  termsS9?: string;
  termsS10?: string;
  termsS11?: string;
  termsS12?: string;
  privacyLastUpdated?: string;
  privacyS1?: string;
  privacyS2?: string;
  privacyS3?: string;
  privacyS4?: string;
  privacyS5?: string;
  privacyS6?: string;
  privacyS7?: string;
  privacyS8?: string;
  privacyS9?: string;
  privacyS10?: string;
  privacyS11?: string;
}

export interface Category {
  id: number;
  name: string;
  description: string;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  description?: string;
}

export interface DeliveryFee {
  id: number;
  state: string;
  feeKobo: number;
}

export interface DeliveryFeeInput {
  state: string;
  feeKobo: number;
}

export interface ContactFormInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
}

export type ListProductsParams = {
flavor?: string;
type?: string;
inStock?: boolean;
categoryId?: number;
search?: string;
};

export type ListOrdersParams = {
status?: string;
limit?: number;
};

export type GetOrdersByEmailParams = {
email: string;
};

export type AdminLogout200 = {
  message: string;
};

export type ChangeAdminPassword200 = {
  message: string;
};

export type SubmitContactForm200 = {
  message: string;
};