# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FFG Foods e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: flavor
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: inStock
          in: query
          required: false
          schema:
            type: boolean
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/by-email:
    get:
      operationId: getOrdersByEmail
      tags: [orders]
      summary: Get orders by customer email
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orders for this email
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
          description: Price in kobo (1 Naira = 100 kobo)
        flavor:
          type: string
          description: "original | hibiscus | ginger-lemon | cinnamon-lemon"
        type:
          type: string
          description: "sachet | box"
        imageUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        stockCount:
          type: integer
        featured:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - priceKobo
        - flavor
        - type
        - inStock
        - stockCount
        - featured
        - createdAt
    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
        flavor:
          type: string
        type:
          type: string
        imageUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        stockCount:
          type: integer
        featured:
          type: boolean
        categoryId:
          type: ["integer", "null"]
      required:
        - name
        - description
        - priceKobo
        - flavor
        - type
        - inStock
        - stockCount
        - featured
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
        flavor:
          type: string
        type:
          type: string
        imageUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        stockCount:
          type: integer
        featured:
          type: boolean
        categoryId:
          type: ["integer", "null"]
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        deliveryState:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotalKobo:
          type: integer
        deliveryFeeKobo:
          type: integer
        totalKobo:
          type: integer
        status:
          type: string
          description: "pending | processing | shipped | delivered | cancelled"
        notes:
          type: ["string", "null"]
        paymentReference:
          type: ["string", "null"]
          description: Paystack payment reference
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - deliveryAddress
        - items
        - subtotalKobo
        - deliveryFeeKobo
        - totalKobo
        - status
        - createdAt
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productType:
          type: string
        quantity:
          type: integer
        priceKobo:
          type: integer
      required:
        - productId
        - productName
        - productType
        - quantity
        - priceKobo
    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        deliveryState:
          type: string
          description: Nigerian state for delivery fee calculation
        notes:
          type: ["string", "null"]
        paymentReference:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerName
        - customerEmail
        - customerPhone
        - deliveryAddress
        - deliveryState
        - items
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - createdAt
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
...