# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FFG Foods e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  name: string;
  description: string;
  createdAt: string;
}