# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FFG Foods e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  description?: string;
}