# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FFG Foods e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress: string;
  /** @nullable */
  deliveryState?: string | null;
  items: OrderItem[];
  subtotalKobo: number;
  deliveryFeeKobo: number;
  totalKobo: number;
  /** pending | processing | shipped | delivered | cancelled */
  status: string;
  /** @nullable */
  notes?: string | null;
  /**
     * Paystack payment reference
     * @nullable
     */
  paymentReference?: string | null;
  createdAt: string;
}