# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FFG Foods e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress: string;
  /** Nigerian state for delivery fee calculation */
  deliveryState: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paymentReference?: string | null;
  items: OrderItemInput[];
}