# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FFG Foods e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  priceKobo: number;
  flavor: string;
  type: string;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  stockCount: number;
  featured: boolean;
  /** @nullable */
  categoryId?: number | null;
}